requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.21.0
requests>=2.28.0
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
pillow>=9.0.0
streamlit>=1.28.0
facebook-business>=17.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
websockets>=11.0.0
